chore!: adaptation on formation calculation due to change on position calculator and minor fixes
resolves #24

+++ formation-calculator-service/src/main/java/br/com/dgc/fmtools/formation_calculator_service/clients/PositionExternal.java
@@ -1,4 +1,5 @@
 package br.com.dgc.fmtools.formation_calculator_service.clients;
 
 public record PositionExternal(
-    String name, int percentage, int quantityKeyAttributes, int quantityPreferableAttributes) {}
+        String name, Double percentage, int quantityKeyAttributes, int quantityPreferableAttributes) {
+}
