+++ formation-calculator-service/src/main/java/br/com/dgc/fmtools/formation_calculator_service/domain/model/positions/GenericPosition.java
@@ -1,11 +1,12 @@
 package br.com.dgc.fmtools.formation_calculator_service.domain.model.positions;
 
-import br.com.dgc.fmtools.formation_calculator_service.domain.model.Position;
 import java.util.UUID;
+
+import br.com.dgc.fmtools.formation_calculator_service.domain.model.Position;
 
 public class GenericPosition extends Position {
 
-  public GenericPosition(String name, UUID playerId, String playerName, int percentage) {
+  public GenericPosition(String name, UUID playerId, String playerName, Double percentage) {
     super(name, playerId, playerName, percentage);
   }
 }
